+++ package.json
@@ -7,8 +7,8 @@
     "start": "cross-env NODE_ENV=production babel-node server/app.js",
     "start:dev": "cross-env NODE_ENV=development nodemon --exec babel-node -- server/app.js",
     "test": "npm run test:server",
-    "pretest:server": "cd server & sequelize --env test db:migrate && sequelize --env test db:seed:all",
-    "cleanup": "cd server & sequelize --env test db:seed:undo:all",
+    "pretest:server": "sequelize --env test db:migrate && sequelize --env test db:seed:all",
+    "cleanup": "sequelize --env test db:seed:undo:all",
     "test:server": "cross-env NODE_ENV=test nyc mocha server/tests",
     "posttest:server": "npm run cleanup",
     "lint:server": "eslint \"server/**/*.js\" app.js --fix",
@@ -26,20 +26,6 @@
   },
   "homepage": "https://github.com/otseobande/Book-A-Meal#readme",
   "dependencies": {
-    "bcrypt": "^2.0.1",
-    "dotenv": "^5.0.1",
-    "express": "^4.16.3",
-    "express-validation": "^1.0.2",
-    "joi": "^13.2.0",
-    "jsonwebtoken": "^8.2.1",
-    "moment": "^2.22.1",
-    "morgan": "^1.9.0",
-    "pg": "^7.4.1",
-    "pg-hstore": "^2.3.2",
-    "sequelize": "^4.37.6",
-    "sinon": "^4.5.0",
-    "sinon-chai": "^3.0.0",
-    "sinon-express-mock": "^2.0.0",
     "babel": "^6.23.0",
     "babel-cli": "^6.26.0",
     "babel-core": "^6.26.3",
@@ -48,23 +34,36 @@
     "babel-plugin-transform-object-rest-spread": "^6.26.0",
     "babel-plugin-transform-runtime": "^6.23.0",
     "babel-preset-env": "^1.6.1",
+    "bcrypt": "^2.0.1",
     "chai": "^4.1.2",
     "chai-http": "^4.0.0",
     "concurrently": "^3.5.1",
     "coveralls": "^3.0.0",
     "cross-env": "^5.1.4",
+    "dotenv": "^5.0.1",
     "eslint": "^4.19.1",
     "eslint-config-airbnb": "^16.1.0",
     "eslint-plugin-import": "^2.11.0",
     "eslint-plugin-jsx-a11y": "^6.0.3",
     "eslint-plugin-react": "^7.7.0",
+    "express": "^4.16.3",
+    "express-group-routes": "^1.1.0",
+    "express-validation": "^1.0.2",
     "istanbul": "^0.4.5",
+    "joi": "^13.2.0",
+    "jsonwebtoken": "^8.2.1",
     "mocha": "^5.1.1",
     "mocha-lcov-reporter": "^1.3.0",
+    "moment": "^2.22.1",
+    "morgan": "^1.9.0",
     "nyc": "^11.7.1",
-    "sequelize-cli": "^4.0.0"
+    "pg": "^7.4.1",
+    "pg-hstore": "^2.3.2",
+    "sequelize": "^4.37.6",
+    "sequelize-cli": "^4.0.0",
+    "sinon": "^4.5.0",
+    "sinon-chai": "^3.0.0",
+    "sinon-express-mock": "^2.0.0"
   },
-  "devDependencies": {
-    
-  }
+  "devDependencies": {}
 }
