--- server/.sequelizerc
@@ -1,8 +1,0 @@
-const path = require('path');
-
-module.exports = {
-  'config': path.resolve('config', 'babelHook.js'),
-  'migrations-path': path.resolve('migrations'),
-  'seeders-path': path.resolve('seeders'),
-  'models-path': path.resolve('models')
-};
