+++ server/app.js
@@ -9,20 +9,15 @@ import apiRoutes from './routes/api';
 import otherRoutes from './routes/others';
 
 const app = express();
-
 const port = process.env.PORT || 3000;
 
 app.use(logger('dev'));
-
 app.use(express.urlencoded({ extended: true }));
 app.use(express.json());
-app.use(trimStrings);
-app.use(apiRoutes);
-app.use(otherRoutes);
-app.use(handleErrors);
+app.use(trimStrings, apiRoutes, otherRoutes, handleErrors);
 
 const server = app.listen(port, () => {
-  console.log(`Server started on http://${server.address().address}:${server.address().port}`);
+  console.log(`Server started on port ${server.address().port}`);
 });
 
 export default app;
