+++ server/routes/api.js
@@ -1,8 +1,12 @@
 import { Router } from 'express';
-import v1Routes from './v1';
+import authRoutes from './auth';
+import mealRoutes from './meals';
+import menuRoutes from './menu';
+import orderRoutes from './orders';
 
 const router = Router();
 
-router.use('/api', v1Routes);
+
+router.use('/api/v1', authRoutes, orderRoutes, menuRoutes, mealRoutes);
 
 export default router;
