--- server/routes/v1/menu.js
@@ -1,21 +1,0 @@
-import { Router } from 'express';
-import MenuController from '../../controllers/menuController';
-import {
-  validateCreate,
-  validateUpdate,
-  validateDate
-} from '../../middlewares/validators/menu';
-import { guard, authorize } from '../../middlewares';
-
-const router = Router();
-
-router.use(authorize);
-router.get('/', MenuController.getSpecificDayMenu);
-router.get('/:date', validateDate, MenuController.getSpecificDayMenu);
-
-router.use(guard('caterer'));
-router.post('/', validateCreate, MenuController.createMenu);
-router.put('/:date', validateUpdate, MenuController.updateMenu);
-router.delete('/:date', validateDate, MenuController.deleteMenu);
-
-export default router;
